LayoutTests/inspector/protocol/database-agent.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/protocol-test.js"></script>
<script>

function test()
{
   RuntimeAgent.evaluate("openDatabase('db1', '1.0', 'Database Agent Test database', 1000);", "database-agent-test", false, theTest);

    function theTest(error, result) {
        var objectId = JSON.parse(result.objectId);
        var testSuite = [
            ["DatabaseAgent", "executeSQL", objectId.id, "CREATE TABLE Table1 (TableID INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT);"],
            ["DatabaseAgent", "executeSQL", objectId.id, "CREATE TABLE Table1 (TableID INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT);"],
            ["DatabaseAgent", "getDatabaseTableNames", objectId.id],
            ["RuntimeAgent", "evaluate", "openDatabase('db2', '1.0', 'Database Agent Test second database', 1000);", "database-agent-test", false],
        ];

        InspectorTest.runProtocolTestSuite("DatabaseAgent", testSuite, ["id", "requestId", "databaseId", "transactionId", "objectId", "code", "message"]);
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Protocol stability test. It is dumping request/response pairs of DatabaseAgent functions.
</p>

</body>
</html>
