LayoutTests/inspector/protocol/database-agent-expected.txt

Protocol stability test. It is dumping request/response pairs of DatabaseAgent functions.

-----------------------------------------------------------
DatabaseAgent.executeSQL(<number>,"CREATE TABLE Table1 (TableID INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT);")

request:
{
    id : <number>
    domain : "Database"
    command : "executeSQL"
    arguments : {
        databaseId : <number>
        query : "CREATE TABLE Table1 (TableID INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT);"
    }
}

response:
{
    requestId : <number>
    body : {
        success : false
        transactionId : <number>
    }
}

-----------------------------------------------------------
DatabaseAgent.executeSQL(<number>,"CREATE TABLE Table1 (TableID INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT);")

request:
{
    id : <number>
    domain : "Database"
    command : "executeSQL"
    arguments : {
        databaseId : <number>
        query : "CREATE TABLE Table1 (TableID INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT);"
    }
}

response:
{
    requestId : <number>
    body : {
        success : false
        transactionId : <number>
    }
}

-----------------------------------------------------------
DatabaseAgent.getDatabaseTableNames(<number>)

request:
{
    id : <number>
    domain : "Database"
    command : "getDatabaseTableNames"
    arguments : {
        databaseId : <number>
    }
}

response:
{
    requestId : <number>
    body : {
        tableNames : {
        }
    }
}

-----------------------------------------------------------
RuntimeAgent.evaluate("openDatabase('db2', '1.0', 'Database Agent Test second database', 1000);","database-agent-test",false)

request:
{
    id : <number>
    domain : "Runtime"
    command : "evaluate"
    arguments : {
        expression : "openDatabase('db2', '1.0', 'Database Agent Test second database', 1000);"
        objectGroup : "database-agent-test"
        includeCommandLineAPI : false
    }
}

event DatabaseAgent.DatabaseOpened
{
    type : "event"
    domain : "Database"
    event : "DatabaseOpened"
    body : {
        database : {
            id : <number>
            domain : ""
            name : "db2"
            version : "1.0"
        }
    }
}

response:
{
    requestId : <number>
    body : {
        result : {
            objectId : <string>
            hasChildren : true
            type : "object"
            description : "Database"
        }
    }
}

===========================================================
Coverage for DatabaseAgent
{
    getDatabaseTableNames : "checked"
    executeSQL : "checked"
    DatabaseOpened : "checked"
    sqlTransactionSucceeded : "not checked"
    sqlTransactionFailed : "not checked"
}
